Wrap the syslog octet decoder for general use
