+++ src/sources/util/unix_stream.rs
@@ -30,10 +30,10 @@ pub fn build_unix_stream_source<D>(
     host_key: String,
     shutdown: ShutdownSignal,
     out: Pipeline,
-    build_event: impl Fn(&str, Option<Bytes>, &str) -> Option<Event> + Clone + Send + Sync + 'static,
+    build_event: impl Fn(&str, Option<Bytes>, Bytes) -> Option<Event> + Clone + Send + Sync + 'static,
 ) -> Source
 where
-    D: Decoder<Item = String> + Clone + Send + 'static,
+    D: Decoder<Item = Bytes> + Clone + Send + 'static,
     D::Error: From<std::io::Error> + std::fmt::Debug + std::fmt::Display,
 {
     let out = out.sink_map_err(|error| error!(message = "Error sending line.", %error));
@@ -76,7 +76,7 @@ where
             let stream = socket.allow_read_until(shutdown.clone().map(|_| ()));
             let mut stream = FramedRead::new(stream, decoder.clone()).filter_map(move |line| {
                 ready(match line {
-                    Ok(line) => build_event(&host_key, received_from.clone(), &line).map(Ok),
+                    Ok(line) => build_event(&host_key, received_from.clone(), line).map(Ok),
                     Err(error) => {
                         emit!(UnixSocketError {
                             error,
